AddContact hooks
