+++ src/screens/WalletScreen/Modals/SendModal/screens/AddContact/AddContact.tsx
@@ -1,44 +1,15 @@
-import React, { useEffect } from 'react';
+import React from 'react';
 import { View } from 'react-native';
 import { Text, Button, Input } from '@components';
-import { useState } from '@hookstate/core';
 import KeyboardSpacer from 'react-native-keyboard-spacer';
-import { isAddress } from 'ethers/lib/utils';
-import { resolveENSAddress } from '@src/model/wallet';
-import { contactCreate } from '@models/contact';
-import { globalContactState } from '@src/stores/ContactStore';
 import styles from './AddContact.styles';
 import { AddContactProps } from './AddContact.types';
+import { useAddContact } from './AddContact.hooks';
 
 const AddContact: React.FC<AddContactProps> = ({ onContactAdded }) => {
-	const state = useState(globalContactState());
-	const [name, setName] = React.useState<string>();
-	const [address, setAddress] = React.useState<string>();
-	const [ensAddress, setEnsAddress] = React.useState<string>();
-
-	useEffect(() => {
-		const lookForENS = async () => {
-			if (address && address.includes('.')) {
-				const resolvedAddress = await resolveENSAddress(address);
-				setEnsAddress(resolvedAddress!);
-			} else {
-				setEnsAddress(undefined);
-			}
-		};
-		lookForENS();
-	}, [address]);
-
-	const validAddress = !!address && (isAddress(address) || isAddress(ensAddress || ''));
-
-	const onContactCreate = async () => {
-		if (name && validAddress) {
-			const newContact = await contactCreate(name, address);
-			if (newContact) {
-				state.contactList.merge([newContact]);
-				onContactAdded();
-			}
-		}
-	};
+	const { name, setName, address, setAddress, ensAddress, validAddress, onContactCreate } = useAddContact({
+		onContactAdded
+	});
 
 	return (
 		<View style={styles.container}>
